Add files via upload
this version is error prone as to my test but not leak tested so do check for valgrind and test with other test cases as well as eating chicken in the mean time, the previous error was the extra space included in write func. which gave in a null char into the results make it mess witht he output while siphoning it into the checker
link for tester : https://github.com/anolivei/Push_swap42, this is the tester and this is how u use the visualizer : /usr/bin/python3 python_visualizer.py 4 23 5 -91292138 8721821 98 454 64584 648448, the numbers being the test case!, and random value tester : !!! given in the github page!, for an amazing checker use dis link : https://github.com/kurval/42-push_swap : type the test case exactly as shown there!